% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_snap.R
\name{aoi_snap}
\alias{aoi_snap}
\title{Snap Area of Interest}
\usage{
aoi_snap(aoi, method = c("expand", "shrink"), buffer = 0)
}
\arguments{
\item{aoi}{is a sf object (e.g. polygon). The bounding box of the shape will
be used to create rectangular shape.}

\item{method}{Options are \emph{shrink} or \emph{expand}. \emph{Shrink} will snap the aoi in
to the nearest 100m. \emph{Expand} will snap the AOI out to the nearest 100m.}

\item{buffer}{adds additional buffer to expand AOI bounding box}
}
\value{
a rectangular sf polygon
}
\description{
Adjusts the area of interest to the nearest 100m.
Note that this package assumes to data is in a metric equal area projection.
}
\details{
This is an essential first step.  As subsequent co-variate layers will be
generated at multiple resolutions (e.g. 5, 10, 25m^2) and then
disaggregate'd back to the lowest resolution.
Having the AOI set 100m break-points facilitates this.
}
\examples{
## Load sf object
file <- system.file("extdata/datecreek_aoi.gpkg", package = "PEMprepr")
aoi_raw <- sf::st_read(dsn = file, quiet = TRUE)
## snap aoi to nearest 100m
aoi_snap(aoi_raw)

}
