% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_snap_landscape.R
\name{aoi_snap_landscape}
\alias{aoi_snap_landscape}
\title{AOI Snap to landscape}
\usage{
aoi_snap_landscape(in_aoi)
}
\arguments{
\item{in_aoi}{}
}
\value{
sf polygon
}
\description{
Creates a watershed based area of interest.  Does this by querying \emph{BCData's}
\emph{Freshwater Atlas}
}
\examples{
aoi_snap_landscape(aoi) ### add more

}
