% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_snap_watershed.R
\name{aoi_snap_watershed}
\alias{aoi_snap_watershed}
\title{AOI Snap to landscape}
\usage{
aoi_snap_watershed(in_aoi, preview = TRUE)
}
\arguments{
\item{in_aoi}{\code{sf} polygon as the Area of Interest. Needs to be a meters based projection (e.g. Albers, UTM)}

\item{preview}{Boolean.  \code{TRUE} will provide a mapview preview of the resulting aoi, the input aoi, and the Freshwater Atlas Watersheds that intersect.}
}
\value{
sf polygon
}
\description{
Creates a watershed based area of interest.  Does this by querying \emph{BCData's}
\emph{Freshwater Atlas}
}
\examples{
aoi_snap_landscape(aoi) ### add more
}
