% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output.R
\name{check_output}
\alias{check_output}
\title{Checks for an output file.}
\usage{
check_output(f.ids, vari.name)
}
\arguments{
\item{f.ids}{list of folder IDS}

\item{vari.name}{the standard name of the variable}
}
\value{
a list of folder IDs with relative and absolute paths
}
\description{
This function is almost exclusively called internally within other PEMprepr functions. It checks the function output against a list of standardized outputs and write it to a specified place on the disk.
}
\examples{
check_output(f.ids = f.ids, vari.name = "aoi.snapped")
[1] TRUE
}
