% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_clean_vectors.R
\name{copy_clean_vectors}
\alias{copy_clean_vectors}
\title{copy_clean_vectors}
\usage{
copy_clean_vectors(
  origindir = fid$shape_dir_0010[1],
  targetdir = fid$shape_dir_1010[1]
)
}
\arguments{
\item{origindir}{text filepath of raw vectors}

\item{targetdir}{text filepath of clean vectors}
}
\value{
TRUE
}
\description{
copy_clean_vectors
}
\examples{
copy_clean_vectors(origindir = fid$shape_dir_0010[1],targetdir = fid$shape_dir_1010[1])
}
