% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aoi.R
\name{create_aoi}
\alias{create_aoi}
\title{Title}
\usage{
create_aoi(
  aoi_dir = file.path(fid$shape_dir_0010[2]),
  out_dir = file.path(fid$shape_dir_1010[1]),
  aoi_vec = "aoi.gpkg"
)
}
\arguments{
\item{aoi_vec}{}
}
\description{
Title
}
