% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_base_vectors.R
\name{create_base_vectors}
\alias{create_base_vectors}
\title{Generates base information vectors for sample planning and modelling using bcdata package}
\usage{
create_base_vectors(
  aoi = fs::path(PEMprepr::read_fid()$dir_1010_vector$path_abs, "aoi_snapped.gpkg"),
  out_dir = PEMprepr::read_fid()$dir_1010_vector$path_abs
)
}
\arguments{
\item{aoi}{An  \code{sf} object (e.g. polygon), which has been snapped to a common extend.
This is commonly the output of the snap_aoi() function. A default location and name are
applied in line with standard workflow.}

\item{out_dir}{A character string of filepath which points to output location. A default
location and name are applied in line with standard workflow.}
}
\value{
multiple gpkg files written directly to out_put directory
}
\description{
Generates base information vectors for sample planning and modelling using bcdata package
}
\examples{
#' ## Load snapped aoi object
aoi_file <- system.file("extdata/datecreek_aoi.gpkg", package = "PEMprepr")
sn_aoi <- snap_aoi(aoi_file, write_output = FALSE)
create_base_vectors(sn_aoi, out_dir = PEMprepr::read_fid()$dir_1010_vector$path_abs)

}
