% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bgc_template.R
\name{create_bgc_template}
\alias{create_bgc_template}
\title{Create BGC template}
\usage{
create_bgc_template(
  bec = fs::path(PEMprepr::read_fid()$dir_1010_vector$path_abs, "bec.gpkg"),
  field = "MAP_LABEL",
  template_rast = NULL,
  write_output = TRUE,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel
)
}
\arguments{
\item{bec}{An \code{sf} object (e.g. polygon) or path to a spatial file of BEC within your AOI.
Usually generated when calling \code{\link[=create_base_vectors]{create_base_vectors()}}.}

\item{field}{field within bec object which defines the BEC units. Default is \code{"MAP_LABEL"}}

\item{template_rast}{A \code{SpatRast} template or path to a spatRast file, which will form the
template resolution and extent output. The template is currently created using the
create_base_raster() function.}

\item{write_output}{should the bec raster be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} under the appropriate resolution subfolder.}

\item{out_dir}{A character string of filepath which points to output location. A default
location and name are applied in line with standard workflow.}
}
\value{
write out bgc raster and return object
}
\description{
Create BGC template
}
\examples{
\dontrun{
bgc_raster = create_bgc_template(
  bec = fs::path(PEMprepr::read_fid()$dir_1010_vector$path_abs, "bec.gpkg"),
  field = "MAP_LABEL",
  template_rast = NULL,
  write_output = TRUE,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel
}
}
