% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates.R
\name{create_covariates}
\alias{create_covariates}
\title{Covariate SAGA raster covarates from base digital terrain dataset.}
\usage{
create_covariates(
  dtm = NULL,
  saga_path = NULL,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel,
  layers = "all"
)
}
\arguments{
\item{dtm}{a character or file path location to raster with elevation data.
Outputs will be produced at the same resolution and extent as input dtm}

\item{saga_path}{a character of the file to the SAGA directory on the analysts system.
On linux systems with SAGA GIS installed Use \code{SAGApath = ""}}

\item{out_dir}{file.path. Directory where covariates will be written. If not
specified uses the default from the \code{fid} folder structure. If files already
exists they will NOT be overwritten.}

\item{layers}{character vector. Covariates to be created. Default is \code{"all"}
\itemize{
\item call \code{PEMprepr::layer_options} for full list.
}}
}
\value{
TRUE
}
\description{
#' Takes a dem and via SAGA GIS generates raster covariates at the resolution of
the dem.
}
\details{
Depending on your system the path to \code{saga_cmd} may need to be specified.
}
\examples{
\dontrun{
#  # start testing
#dtm = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel,"25m","dem.tif")
#  saga_path = "C:\\Programs\\saga-9.2.0_x64\\saga-9.2.0_x64"
#  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel
#  layers = "all"
#  # end testing
#--- load in dem and write to tempfile ---#
aoi_raw <- system.file("extdata", "aoi.gpkg", package ="PEMprepr")
aoi_raw <- sf::st_read(aoi_raw)
aoi <- PEMprepr::aoi_snap(aoi_raw, "shrink")
t25 <- create_template(aoi, 25)
trim_raw <- cded_raster(aoi)
trim <- terra::rast(trim_raw)
dtm <- terra::project(trim, t25)
tmp <- tempfile(fileext = ".tif")
writeRaster(dtm, tmp)


#--- create all SAGA covariates ---#
create_covariates(dtm = tmp, layers = "all", output = dir, SAGApath = SAGApath)
}
}
