% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fine_res.R
\name{create_fine_res}
\alias{create_fine_res}
\title{Create high resolution versions of rasters}
\usage{
create_fine_res(
  inputFileList,
  output = "./10_clean_inputs/covariates/5",
  targetRes = 5
)
}
\arguments{
\item{inputFileList}{a character vector specifying the location of all the input rasters.  Input list should only be a raster type (e.g. .tif, .sdat).  Best practice is to use list.files(full.names = TRUE).}

\item{output}{destination of the output files. \emph{NOTE} files names will have a suffix added to indicate the original resolution.}

\item{targetRes}{desired resolution to convert to.}
}
\description{
Uses \code{terra::disagg} to push a set of rasters to a target resolution.  Note that input raster resolution need to be divisible by the target resolution.
}
\examples{
l <- list.files("e:/covariates/10")
create_fine_res(l, output = "e:/covariates/5", targetRes = 5)
}
\keyword{disaggregate}
