% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fuzzy_matrix.R
\name{create_fuzzy_matrix}
\alias{create_fuzzy_matrix}
\title{Create a fuzzy matrix}
\usage{
create_fuzzy_matrix(map_unit, edatopic, neighbours)
}
\arguments{
\item{map_unit}{Single BGC to create fuzzy for}

\item{edatopic}{Data.table of edatopic data}

\item{neighbours}{data.table of neighbours}
}
\value{
Matrix of fuzziness
}
\description{
Create a fuzzy matrix
}
\examples{
create_fuzzy_matrix(map_unit, edatopic, neighbours )
}
