% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_landscape_covariates.R
\name{create_samplr_covariates}
\alias{create_samplr_covariates}
\title{Generate digital terrain model derived covariates for sample plan development using SAGA GIS.}
\usage{
create_samplr_covariates(
  dtm = dtm,
  saga_path = NULL,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel,
  layers = c("mrvbf", "dah", "landform"),
  sieve_size = 10,
  dah_threshold = 0.2,
  saga_param = list(T_SLOPE = 64, TPCTL_V = 6, T_PCTL_R = 2, P_SLOPE = 4, P_PCTL = 3,
    UPDATE = 1, CLASSIFY = 1, MAX_RES = 100)
)
}
\arguments{
\item{dtm}{An \code{SpatRast} object or path to a spatial file (.tif) with digital elevation data.
Should be a meter based coordinate reference system. location to raster with elevation data.
Outputs layers will be produced at the same resolution and extent as input dtm}

\item{saga_path}{saga_path a \code{character} of the file to the SAGA directory on the analysts system.}

\item{out_dir}{A \code{character} string of filepath which points to output location. A default
location and name are applied in line with standard workflow.}

\item{layers}{A \code{character} vector describing the covariates to be created. By default all three will
be generates \code{c("mrvbf", "dah", "landform")}.}

\item{sieve_size}{a \code{numeric}value to define the number of pixels to be grouped during the sieving process.
default is \code{10}.}

\item{dah_threshold}{a \code{numeric}value to define threshold at which diurnal anisotrophic heating (dah) will
define catergories. Default value is \code{10}.}

\item{saga_param}{named list. Parameters that are fed to SAGA MRVBF function. IF supplied it must contain
all parameters specified in the SAGA command.}
}
\value{
path to the output directory where files are written (invisibly).
}
\description{
Generate digital terrain model derived covariates for sample plan development using SAGA GIS.
}
\examples{
\dontrun{
#  # start testing
#dtm = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel,"25m","dem.tif")
#  saga_path = "C:\\Programs\\saga-9.2.0_x64\\saga-9.2.0_x64"
#  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel
#  layers = "all"
#  # end testing
#--- load in dem and write to tempfile ---#
aoi_raw <- system.file("extdata", "aoi.gpkg", package ="PEMprepr")
aoi_raw <- sf::st_read(aoi_raw)
aoi <- PEMprepr::aoi_snap(aoi_raw, "shrink")
t25 <- create_template(aoi, 25)
trim_raw <- cded_raster(aoi)
trim <- terra::rast(trim_raw)
dtm <- terra::project(trim, t25)
tmp <- tempfile(fileext = ".tif")
writeRaster(dtm, tmp)


#--- create all SAGA covariates ---#
create_samplr_covariate(dtm = tmp, saga_path = saga_path,
out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel )
}
}
