% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{find_saga_path}
\alias{find_saga_path}
\title{Discover the path to your SAGA installation (\code{saga_cmd(.exe)} executable)}
\usage{
find_saga_path(root = NULL)
}
\arguments{
\item{root}{where to start the search. The defaults are:
\itemize{
\item Windows: "C:/"
\item MacOS: "/Applications"
\item Linux: "/usr"
}}
}
\value{
path to \code{saga_cmd} executable
}
\description{
Searches in common places where SAGA could be installed. If you know where it
is, supply the path to \code{root}. It also checks to see if there is a \code{saga_cmd}
on the \code{PATH}. After the installation is found, it validates it and checks
for an adequate version via \code{\link[=saga_cmd]{saga_cmd()}}.
}
\details{
If you don't have SAGA installed:
\itemize{
\item On Windows or MacOS, install the latest version from:
\url{https://sourceforge.net/projects/saga-gis/files/}. On MacOS, it is
recommended to move \code{SAGA.app} into your \verb{/Applications} folder, and you may
need to turn on permissions for the SAGA in your System Preferences ->
Privacy and Security.
\item Linux: On Ubuntu and Debian run \verb{apt install saga}
}
}
\seealso{
\code{\link[=saga_cmd]{saga_cmd()}}
}
