% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cded_dem.R
\name{get_cded_dem}
\alias{get_cded_dem}
\title{#' Get Canadian Digital Elevation Data (CDED) to match area of interest}
\usage{
get_cded_dem(
  aoi = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_abs, "25m",
    "template.tif"),
  res = FALSE,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_abs,
  write_output = TRUE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{aoi}{An \code{SpatRast} object or path to a spatial file (.tif) usually generated
when calling \code{\link[=create_template_raster]{create_template_raster()}}. Should be a meter based coordinate reference system.}

\item{res}{resolution in meters of final project. If no values specified (default is FALSE),
output spatRast will match the resolution of aoi.}

\item{out_dir}{the root directory to hold the cded dem spatRast file. If not
specified uses the default from the \code{fid} folder structure.}

\item{write_output}{should the cded spatRast be written to disk?
If \code{TRUE} (default), will write to \code{out_dir} in a subfolder defined by resolution e.g: "25m".}

\item{overwrite}{logical. If \code{TRUE}, \code{filename} is overwritten}

\item{...}{arguments passed on to \code{\link[bcmaps:cded_terra]{bcmaps::cded_terra()}}}
}
\value{
a spatRast file of CDED data
}
\description{
This function is a wrapper for bcmaps::cded_terra() to extract data from the
Canadian Digital Elevation Data set to match the raster template built using
create_raster_template().
}
\examples{
\dontrun{
get_cded_dem(
  aoi = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_abs, "25m", "template.tif"),
  res = FALSE,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_abs,
  write_output = TRUE
  overwrite = FALSE)
}
}
