% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roads.R
\name{get_roads}
\alias{get_roads}
\title{#' Generates road layers(roads and forest service roads) in vector format sample planning and modelling using bcdata package}
\usage{
get_roads(aoi, out_dir)
}
\arguments{
\item{aoi}{An \code{sf} object (e.g. polygon) or path to a spatial file,
which has been snapped to a common extend.
This is commonly the output of the snap_aoi() function. A default location and name are
applied in line with standard workflow.}

\item{out_dir}{A character string of filepath which points to output location. A default
location and name are applied in line with standard workflow.}
}
\value{
writes out a \code{sf} object with roads information for the given aoi as "road_network.gpkg"
}
\description{
#' Generates road layers(roads and forest service roads) in vector format sample planning and modelling using bcdata package
}
\examples{
\dontrun{
aoi_file <- system.file("extdata/datecreek_aoi.gpkg", package = "PEMprepr")
roads <- get_roads(aoi_file, out_dir = PEMprepr::read_fid()$dir_1010_vector$path_abs)
}
}
