% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trim_dem.R
\name{get_trim_dem}
\alias{get_trim_dem}
\title{Get Trim data form Cded}
\usage{
get_trim_dem(
  aoi_bb = file.path(fid$shape_dir_1010[2], "aoi_snapped.gpkg"),
  res,
  out_dir = fid$cov_dir_1020[2]
)
}
\arguments{
\item{aoi_bb}{filepath to aoi required}

\item{res}{resolution of raster}

\item{out_dir}{output drive, default is set to fid$cov_dir_1020\link{2}}
}
\value{
SpatRast DEM from CDED
}
\description{
Get Trim data form Cded
}
\examples{
get_trim_dem(aoi_bb = file.path(fid$shape_dir_1010[2],"aoi_snapped.gpkg"), res = 5, out_dir = fid$cov_dir_1020[2]))
}
