% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-layers.R
\name{Layers}
\alias{Layers}
\alias{recursive_layers_call}
\alias{read_crop}
\alias{covariate_file_names}
\title{Layers}
\usage{
recursive_layers_call(layers, moddir = moddir, artifact = artifacts)

read_crop(f, poly, tmp)

covariate_file_names(outputdir, nm)
}
\arguments{
\item{layers}{Character vector. SAGA modules to call}

\item{moddir}{Internal Data. Cartesian coordinates of each strata}

\item{artifact}{Internal Data. Determines proportion of cells to plot}
}
\value{
Vector of recursive layers to call.
}
\description{
Layers

Recursive layers call

Crop tiles

Covariate file paths
}
\concept{Layers}
\keyword{internal}
