% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_covariates.R
\name{mosaic_covariates}
\alias{mosaic_covariates}
\title{Mosaic covariates}
\usage{
mosaic_covariates(dir, output, tmp, poly)
}
\arguments{
\item{dir}{Character. Directory where covaraites exist}

\item{output}{Character. Directory to write mosaiced covaraites}

\item{tmp}{Character. Temp folder directory to be used for mosaicing.
Deleted upon completion of function.}

\item{poly}{spatVector. Original tile boundaries generated using \code{preprocess_dtm}}
}
\description{
Mosaic covariates
}
