% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_res.R
\name{multi_res}
\alias{multi_res}
\title{Multi-resolutions}
\usage{
multi_res(input, output = "CoVars", resolution = c(2.5, 5, 10, 25))
}
\arguments{
\item{input}{a \code{raster::raster} object (not a file)}

\item{output}{location of the out rasters}

\item{resolution}{= c(5, 10, 25)  multiple resolutions can be specified}
}
\description{
Provides multiple resolutions of the input raster
Resamples raster to specified desired resolution using the raster package.
NOTE: an attempt was made to use the SAGA resample but I had trouble aligning the rasters. I am sure that gdal can be used but for now funtioning is internal to R packages.
}
\details{
For the PEM project we generally start with a high resolution dtm.  Resampling to a set of lower resolution DTMs is used to capture multi-scale influences on the ecological unit.
NOTE: to ensure raster stacking of all covariates this operation should be completed only on data that has been cropped to the AOI using aoi_snap()
}
\examples{
## generate default raster sizes
multi_res(r)
}
