% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEMprepr-package.R
\docType{package}
\name{PEMprepr-package}
\alias{PEMprepr}
\alias{PEMprepr-package}
\title{PEMprepr: Prepare project for PEMr analysis}
\description{
Sets up project template and folder structure, and prepares raster and vector layers for use in a PEMr project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ninoxconsulting/PEMprepr}
  \item Report bugs at \url{https://github.com/ninoxconsulting/PEMprepr/issues}
}

}
\author{
\strong{Maintainer}: Genevieve Perkins \email{gperkins@ninoxconsulting.ca} (\href{https://orcid.org/0000-0003-2663-5904}{ORCID})

Authors:
\itemize{
  \item Will McKenzie \email{Will.MacKenzie@gov.bc.ca}
  \item Kiri Daust \email{kiri.daust@gov.bc.ca}
  \item Colin Chisholm \email{Colin.Chisholm@unbc.ca}
  \item Matthew Coghill \email{matthewcoghill@live.ca}
  \item Claire Armour \email{claire.armour@gov.bc.ca}
  \item Andy Teucher \email{andy@andyteucher.ca} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})
}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
