% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dtm.R
\name{preprocess_dtm}
\alias{preprocess_dtm}
\title{Pre-process dtm}
\usage{
preprocess_dtm(
  dtm,
  template = NULL,
  odir,
  tile = FALSE,
  ncol = 5,
  nrow = 10,
  tilebuffer = NULL,
  writeoutputs = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{dtm}{file path or SpatRaster. dtm of interest}

\item{template}{file path or SpatRaster. Template to align dtm}

\item{odir}{file path. Output directory}

\item{tile}{Logical. If \code{TRUE}, \code{dtm} will be tiled. Defaults to \code{FALSE}}

\item{ncol}{Numeric. Number of columns to tile \code{dtm}.}

\item{nrow}{numeric. Number of rows to tile \code{dtm}.}

\item{tilebuffer}{Numeric. Distance (in dtm units) to buffer tiles if desired}

\item{writeoutputs}{Logical. Write outputs to disk if \code{TRUE}}

\item{overwrite}{Logical. Overwrite previous files}
}
\description{
pre-process dtm to use in create_covariates()
}
