% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_output.R
\name{pull_output}
\alias{pull_output}
\title{Reads/Writes function output to user-defined folder set}
\usage{
pull_output(obj, f.ids, vari.name)
}
\arguments{
\item{f.ids}{list of folder IDS}

\item{vari.name}{the standard name of the variable}
}
\value{
a list of folder IDs with relative and absolute paths
}
\description{
This function is almost exclusively called internally within other PEMprepr functions. It checks the function output against a list of standardized outputs and write it to a specified place on the disk.
}
\examples{
pull_output(f.ids = f.ids, vari.name = "aoi.snapped")
[1] TRUE
}
