% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_align.R
\name{raster_align}
\alias{raster_align}
\title{Raster Align}
\usage{
raster_align(raster, polygon)
}
\arguments{
\item{raster}{single channel terra::rast object}

\item{polygon}{the aoi polygon@return something@export  uncomment this to make it usable}
}
\description{
Ensures raster aligns with vector data.
}
\details{
For example, does the DEM align with the output of \code{aoi_snap}
}
