% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{saga_cmd}
\alias{saga_cmd}
\title{Validate and return the path to SAGA \code{saga_cmd(.exe)} executable}
\usage{
saga_cmd(saga_path = saved_saga_path())
}
\arguments{
\item{saga_path}{Path to \code{saga_cmd} executable file. By default checks to
see if it has been set yet this sessions, and if not consults
\code{getOption("pemprepr.saga_path")}.}
}
\value{
path to saga_cmd executable
}
\description{
Accepts a path to the \code{saga_cmd(.exe)} executable, or looks in
\code{getOption("pemprepr.saga_path")}. After being run once, sets an internal
variable so that the path is remembered in subsequent runs in the same R
session.
}
\details{
It is recommended to set the option in your \code{.Rprofile} file with:
\code{options("pemprepr.saga_path" = "path to your saga_cmd(.exe) file")}. This
will ensure SAGA is always found when it is needed.
}
\seealso{
\code{\link[=find_saga_path]{find_saga_path()}}
}
