% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{setup_folders}
\alias{setup_folders}
\title{Setup PEM folder structure}
\usage{
setup_folders(aoi_name)
}
\arguments{
\item{aoi_name}{A character string to name the main data folder}
}
\value{
a named list of the folders created
}
\description{
Sets up the default folder structure for all analysis and modelling
}
\details{
Folder structure contains all the raw data, covariates generated, subsequent models, and predicted map rasters. \emph{Note} if directories have been created this function can be used to generate a list of the folder_id and associated file paths.
}
\examples{
## add more
setup_folders("AppleValley")
}
