% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders2.R
\name{setup_folders2}
\alias{setup_folders2}
\title{Setup PEM folder structure}
\usage{
setup_folders2(aoi.name)
}
\arguments{
\item{aoi.name}{A character string to name the main data folder. This will become the working directory for the entire project}
}
\value{
a named list of the folders created
}
\description{
Sets up the default folder structure for all analysis and modelling.
}
\details{
Folder structure contains all the raw data, covariates generated, subsequent models, and predicted map rasters. \emph{Note} If directories have been created, this function can be used to generate a list of the folder_id and associated file paths.

CLAIRE NOTES:
-Need to ensure that the folder_names.xlsx is properly referenced from the package.
-How can we make sure the absolute paths being created are in the appropriate working folder? Is this set in the pipeline?
}
\examples{
setup_folders2("Kitimat")
}
