% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_folders.R
\name{setupfolders}
\alias{setupfolders}
\title{Setup Folder Structure}
\usage{
setupfolders(aoi)
}
\arguments{
\item{aoi}{A character string to name the main data folder}
}
\value{
a named list of the folders created
}
\description{
Sets up the default folder structure
}
\details{
Folder structure contains all the raw data, covariates generated, and
subsequent models, and predicted map rasters. \emph{Note} if directories have
been created this function can be used to generate a list of the folder_id
and associated file paths.
}
\examples{
## add more
setupfolder("AppleValley")
list.dir
}
