% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_aoi.R
\name{snap_aoi}
\alias{snap_aoi}
\title{Snap Area of Interest}
\usage{
snap_aoi(
  aoi = NULL,
  method = c("expand", "shrink"),
  buffer = 0,
  write_output = TRUE,
  out_dir = read_fid()$dir_1010_vector$path_abs
)
}
\arguments{
\item{aoi}{An `"sf"`` object (e.g. polygon) or a path to a spatial file
representing the area of interest. The bounding box of the aoi will
be used to create a rectangular shape.}

\item{method}{Options are \code{"shrink"} or \code{"expand"}. \code{"shrink"} will snap
the aoi in to the nearest 100m. \code{"Expand"} will snap the AOI out to the
nearest 100m.}

\item{buffer}{Additional buffer to expand AOI bounding box. Only used when
\code{method = "expand"}}

\item{write_output}{should the snapped aoi bounding box be written to disk?
If \code{TRUE} (default), will write to \code{out_dir}}

\item{out_dir}{the directory to hold the snapped boundary file. If not
specified uses the default from the \code{fid} folder structure}
}
\value{
a rectangular sf polygon
}
\description{
Creates a rectangular polygon from the area of interest,
rounded to the nearest 100m.
Note that this package assumes to data is in a metric equal area projection.
}
\details{
This is an essential first step.  As subsequent co-variate layers will be
generated at multiple resolutions (e.g. 5, 10, 25m^2) and then
disaggregate'd back to the lowest resolution.
Having the AOI set 100m break-points facilitates this.
}
\examples{
## Load sf object
aoi_file <- system.file("extdata/datecreek_aoi.gpkg", package = "PEMprepr")
## snap aoi to nearest 100m
snap_aoi(aoi_file, write_output = FALSE)

}
