% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_raster.R
\name{align_raster}
\alias{align_raster}
\title{Align raster}
\usage{
align_raster(raster, rtemplate, outpath = "")
}
\arguments{
\item{raster}{a \code{terra::rast} object}

\item{rtemplate}{a \code{terra::rast} object}

\item{outpath}{a character string to save location}
}
\description{
Forces raster to an specified extent
}
\details{
To facilitate stacking of rasters, and rasters of varied resolutions rasters
must have the same extent. This function ensure that ..... MORE HERE ...
}
\examples{
add some examples

}
