% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aoi.R
\name{create_aoi}
\alias{create_aoi}
\title{Create Area of interest}
\usage{
create_aoi(
  aoi_dir = file.path(fid$shape_dir_0010[2]),
  out_dir = file.path(fid$shape_dir_1010[1]),
  aoi_vec = "aoi.gpkg"
)
}
\arguments{
\item{aoi_dir}{location of the raw aoi}

\item{out_dir}{output location default is}

\item{aoi_vec}{text string of name of raw aoi, default = "aoi.gpkg"}
}
\value{
TRUE
}
\description{
Create Area of interest
}
\examples{
create_aoi(aoi = file.path(fid$shape_dir_0010[2], out_dir = file.path(fid$shape_dir_1010[1],aoi_vec = "aoi.gpkg")
}
