% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aoi.R
\name{create_aoi}
\alias{create_aoi}
\title{Create a rectangular AOI bounding box}
\usage{
create_aoi(
  aoi_dir = read_fid()$dir_0010_vector$path_abs,
  filename = NULL,
  out_dir = read_fid()$dir_1010_vector$path_abs,
  ...
)
}
\arguments{
\item{aoi_dir}{the directory containing the AOI boundary file. If not
specified uses the default from the \code{read_fid()} folder structure}

\item{filename}{the input file name. If not specified, and there is only one
spatial file in \code{aoi_dir}, it will use that.}

\item{out_dir}{the directory to hold the snapped boundary file. If not
specified uses the default from the \code{fid} folder structure}

\item{...}{Arguments passed on to \code{\link[=aoi_snap]{aoi_snap()}}}
}
\value{
path to the snapped aoi file, invisibly
}
\description{
Create a rectangular AOI bounding box containing your AOI, that is snapped
to a 100m grid to facilitate alignment with rasters.
}
\examples{
\dontrun{
create_aoi(filename = "my_aoi.gpkg")
}
}
