% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_base_vectors.R
\name{create_base_vectors}
\alias{create_base_vectors}
\title{Retrieves essential base vectors}
\usage{
create_base_vectors(
  in_aoi = file.path(fid$shape_dir_1010[2], "aoi_snapped.gpkg"),
  out_path = fid$shape_dir_0010[1]
)
}
\arguments{
\item{in_aoi}{\code{sf} polygon}

\item{out_path}{character string that points the output location}
}
\description{
Collects vector data from the \href{https://catalogue.data.gov.bc.ca}{BC Data Catelog} for a spcified area of interest.
}
\details{
This script downloads the relevant spatial data for BEC zones, VRI, TEM,
waterbodies and the road network used in stage 1 of PEM processing.
Data is downloaded directly from the
\href{https://catalogue.data.gov.bc.ca/dataset?download_audience=Public}{BC Data Catalogue}
using the \href{https://github.com/bcgov/bcdata}{bcdata} package.

Datasets retrieved include:
\itemize{
\item BEC - \href{https://catalogue.data.gov.bc.ca/dataset/bec-map}{Biogeoclimatic Ecosystems Classification} are used to define and/or select specific "subzones" within a defined study area. Note this can be also pointed to different file where more updated version is available.
\item Vegetation Resource Inventory (VRI) - This layer includes a variety of vegetation measure, including cutblock age, TEM data and ..... Detailed data standards can be found \href{https://www2.gov.bc.ca/gov/content/industry/forestry/managing-our-forest-resources/forest-inventory/data-management-and-access/vri-data-standards}{here}.
\item Freshwater Atlas -  The atlas is separated into the different types of waterbodies (lakes, rivers, wetlands, streams, man made, etc.), which requires a seperate download per type. Alternatively a single combined layer can be downloaded by is limited to linear data type. A parameter within the function can be set to "polygon" (the default option) or "linear".
\item Road network THIS STILL NEEEDS WORK
\item Road network - In previous works, the \href{https://catalogue.data.gov.bc.ca/dataset/digital-road-atlas-dra-master-partially-attributed-roads}{raw road network} was found to be too detailed. The raw road network is filtered to only keep named roads, and then the up to date \href{https://catalogue.data.gov.bc.ca/dataset/forest-tenure-road-segment-lines}{Forest Service Road (FSR) layer} is downloaded and merged with the filtered road network to produce a really good representation of where roads are actually located on the landscape.
\item Fire - Fire and fire intensity are used used to assist in identifying areas where a high cost is applied to reduce sampling. This includes:
-\href{https://catalogue.data.gov.bc.ca/dataset/fire-perimeters-current}{current}
-\href{https://catalogue.data.gov.bc.ca/dataset/fire-perimeters-historical}{historical} polygons.
-\href{https://catalogue.data.gov.bc.ca/dataset/c58a54e5-76b7-4921-94a7-b5998484e697}{severity}.
\item Accessing the \href{https://catalogue.data.gov.bc.ca/dataset/harvested-areas-of-bc-consolidated-cutblocks-}{consolidated cutblock layer} using the bcdata package. The FTen layer is also used to identify very recent cutblocks.
}

Note the detailed
Tenure (private ownership) 	WHSE_CADASTRE.PMBC_PARCEL_FABRIC_POLY_FA_SVW	BCGW

Detailed roads layers. In some cases regionally specific roads are available through other collaborative projects (cumulative effects). These are not directly available through the BCGW.
}
\examples{
create_base_vectors <- function(in_aoi = your_polygon, out_path = "00_raw_inputs/vector")

}
