% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bgc_template.R
\name{create_bgc_template}
\alias{create_bgc_template}
\title{Create BGC template}
\usage{
create_bgc_template(
  bec_sf,
  field = "MAP_LABEL",
  template = r25,
  outpath = fid$cov_dir_1020[2]
)
}
\arguments{
\item{bec_sf}{sf object with bec linework}

\item{field}{field within in sf object which will be used to assign raster value}

\item{template}{raster object of the size in which to model output raster not including the resoution subfolder}

\item{outpath}{location in which output to be saved, default is based on fid object. Note resolution of template will be used determine folder in which output is saved}
}
\value{
write out bgc raster and return object
}
\description{
Create BGC template
}
\examples{
bgc_raster = create_bgc_template( bec_sf, field = "MAP_LABEL", template = r25,outpath = fid$cov_dir_1020[2])
}
