% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates.R
\name{create_covariates}
\alias{create_covariates}
\title{Covariate raster generation from input dtm.}
\usage{
create_covariates(dtm, SAGApath = "", output = "./cv-rasters", layers = "all")
}
\arguments{
\item{dtm}{file.path Path to raster file.}

\item{SAGApath}{file.path SAGA directory on system.
On linux systems with SAGA GIS installed Use \code{SAGApath = ""}}

\item{output}{file.path Directory where covariates will be written.}

\item{layers}{character vector Covariates to be created. Default is \code(\code{all})
\itemize{
\item call \code(\code{PEMprepr::layer_options}) for full list
}}
}
\description{
\emph{This is an update of cov_dtm.R replacing the processing with 02a} from  BEC_DevExchange_Work_
}
\details{
Takes a dtm and via SAGA GIS generates the covariates embeded in this function.
This script has been tested with SAGA 7.3 on Windows and Ubuntu 18.
Depending on your system the path to \code{saga_cmd} may need to be specified.
}
\examples{
##
create_covariates(dtm,                ## path to dtm file
        SAGApath = "C:/SAGA/"         ## path to saga directory
        output   = "c:/dtm-derived" ) ## path to desired output directory
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
