% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates.R
\name{create_covariates}
\alias{create_covariates}
\title{Covariate raster generation from input dtm.}
\usage{
create_covariates(dtm, SAGApath = "", output = "./cv-rasters", layers = "all")
}
\arguments{
\item{dtm}{file.path. Path to raster file.}

\item{SAGApath}{file.path. SAGA directory on system.
On linux systems with SAGA GIS installed Use \code{SAGApath = ""}}

\item{output}{file.path. Directory where covariates will be written. If files already exist they will NOT be overwritten.}

\item{layers}{character vector. Covariates to be created. Default is \code(\code{all})
\itemize{
\item call \code(\code{PEMprepr::layer_options}) for full list.
}}
}
\description{
\emph{This is an update of cov_dtm.R replacing the processing with 02a} from  BEC_DevExchange_Work_
}
\details{
Takes a dtm and via SAGA GIS generates the covariates embeded in this function.
This script has been tested with SAGA 8.4.1 on Windows.
Depending on your system the path to \code{saga_cmd} may need to be specified.
}
\examples{
\dontrun{
#--- load in dtm and write to tempfile ---#
aoi_raw <- system.file("extdata", "aoi.gpkg", package ="PEMprepr")
aoi_raw <- sf::st_read(aoi_raw)
aoi <- PEMprepr::aoi_snap(aoi_raw, "shrink")
t25 <- create_template(aoi, 25)
trim_raw <- cded_raster(aoi)
trim <- terra::rast(trim_raw)
dtm <- terra::project(trim, t25)
tmp <- tempfile(fileext = ".tif")
writeRaster(dtm, tmp)

#--- global vars ---#
dir <- "" #desired base output folder
SAGApath <- "" #change SAGA path to your own

#--- create all SAGA covariates ---#
create_covariates(dtm = tmp, layers = "all", output = dir, SAGApath = SAGApath)
}

}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
