% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates.R
\name{create_covariates}
\alias{create_covariates}
\title{Covariate raster generation from input dtm.}
\usage{
create_covariates(dtm, SAGApath = "", output = "./cv-rasters", layers = "all")
}
\arguments{
\item{dtm}{is a dtm SpatRaster object}

\item{SAGApath}{Is the location of SAGA on your system.  On linux systems with SAGA GIS installed Use \code{SAGApath = ""}}

\item{output}{Location of where rasters will be saved.}

\item{layers}{The covariates that will be generated.  A full list of covariates is listed at: ADD}
}
\description{
\emph{This is an update of cov_dtm.R replacing the processing with 02a} from  BEC_DevExchange_Work_
}
\details{
Takes a dtm and via SAGA GIS generates the covariates embeded in this function.
This script has been tested with SAGA 7.3 on Windows and Ubuntu 18.
Depending on your system the path to \code{saga_cmd} may need to be specified.
}
\examples{
##
create_covariates(dtm,                ## the dtm (loaded by raster() )
        SAGApath = "C:/SAGA/"         ## specify location of SAGA on your machine
        output   = "c:/dtm-derived" ) ## specify output folder
}
\keyword{SAGA,}
\keyword{covariates,}
\keyword{predictors,}
\keyword{raster}
