% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates.R
\name{create_covariates}
\alias{create_covariates}
\title{Generate digital terrain model derived covariates using SAGA GIS.}
\usage{
create_covariates(
  dtm = NULL,
  saga_path = NULL,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel,
  layers = "all",
  tile = FALSE
)
}
\arguments{
\item{dtm}{An \code{SpatRast} object or path to a spatial file (.tif) with digital elevation data.
Should be a meter based coordinate reference system. location to raster with elevation data.
Outputs layers will be produced at the same resolution and extent as input dtm}

\item{saga_path}{a \code{character} of the file to the SAGA directory on the analysts system.
On linux systems with SAGA GIS installed Use \code{SAGApath = ""}}

\item{out_dir}{file.path. Directory where covariates will be written. If not
specified uses the default from the \code{fid} folder structure. If files already
exists they will NOT be overwritten.}

\item{layers}{character vector. Covariates to be created. Default is \code{"all"}.}

\item{tile}{A \code{logical} to define if the dtm needs to be run in tiles.
Currently placeholder to be developed.}
}
\value{
TRUE
}
\description{
Generates suit of dem derived covariates usind SAGA GIS executable file.
Depending on your system the path to \code{saga_path} may need to be specified.
}
\examples{
\dontrun{
#  # start testing
#dtm = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel,"25m","dem.tif")
#  saga_path = "C:\\Programs\\saga-9.2.0_x64\\saga-9.2.0_x64"
#  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel
#  layers = "all"
#  # end testing
#--- load in dem and write to tempfile ---#
aoi_raw <- system.file("extdata", "aoi.gpkg", package ="PEMprepr")
aoi_raw <- sf::st_read(aoi_raw)
aoi <- PEMprepr::aoi_snap(aoi_raw, "shrink")
t25 <- create_template(aoi, 25)
trim_raw <- cded_raster(aoi)
trim <- terra::rast(trim_raw)
dtm <- terra::project(trim, t25)
tmp <- tempfile(fileext = ".tif")
writeRaster(dtm, tmp)


#--- create all SAGA covariates ---#
create_covariates(dtm = tmp, layers = "all", output = dir, SAGApath = SAGApath)
}
}
