% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates.R
\name{create_covariates}
\alias{create_covariates}
\title{Generate digital terrain model derived covariates using SAGA GIS.}
\usage{
create_covariates(
  dtm = NULL,
  saga_path = NULL,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel,
  layers = "all",
  tile = FALSE
)
}
\arguments{
\item{dtm}{An \code{SpatRast} object or path to a spatial file (.tif) with digital elevation data.
Should be a meter based coordinate reference system. location to raster with elevation data.
Outputs layers will be produced at the same resolution and extent as input dtm}

\item{saga_path}{a \code{character} of the file to the SAGA directory on the analysts system.
Use find_saga_path() to locate the approraite saga installation on your machine.}

\item{out_dir}{afile.path. Directory where covariates will be written. If not
specified uses the default from the \code{fid} folder structure. If files already
exists they will NOT be overwritten.}

\item{layers}{A \code{character} vector. Covariates to be created. Default is \code{"all"}.}

\item{tile}{A \code{logical} to define if the dtm needs to be run in tiles.
Currently placeholder to be developed.}
}
\value{
path to the output directory where files are written (invisibly).
}
\description{
Generates suit of dem derived covariates usind SAGA GIS executable file.
Depending on your system the path to \code{saga_path} may need to be specified.
}
\examples{
\dontrun{
#--- create all SAGA covariates ---#
create_covariates(
  dtm = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_rel,"25m","dem.tif",
  saga_path = find_saga_path[3]
  layers = "all",
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_rel
 )
}
}
