% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{Create a raster template from AOI}
\usage{
create_template(aoi, res, outpath = "./10_clean_inputs/covariates")
}
\arguments{
\item{aoi}{is a sf or terra::vect object bounding box created expanded in aoi_snap function(e.g. polygon).  Should be a meter based projection}

\item{res}{desired resolution of the raster (in meters)}

\item{outpath}{output path.  Note that the results will be placed in a subfolder labelled with the resolution.}
}
\value{
a terra raster
}
\description{
Sets the baseline raster template to align and standardize all raster predictor layers
This package assumes to data is in a metric equal area projection most likely BCAlbers
}
\details{
This is an essential first step.  As subsequent co-variate layers will be generated at multiple resolutions (e.g. 5, 10, 25m^2) and then disaggregate'd back to the lowest resolution.
Having the aoi set 100m break-points facilitates this.
}
\examples{

create_template(aoi, 10)
}
\keyword{align}
\keyword{aoi,}
\keyword{crop,}
\keyword{raster,}
\keyword{template,}
