% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cded_dem.R
\name{get_cded_dem}
\alias{get_cded_dem}
\title{#' Get Canadian Digital Elevation Data (CDED) to match area of interest}
\usage{
get_cded_dem(
  aoi = fs::path(PEMprepr::read_fid()$dir_1020_covariates$path_abs, "25m",
    "template.tif"),
  res = 5,
  out_dir = PEMprepr::read_fid()$dir_1020_covariates$path_abs,
  write_output = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{aoi}{the \code{SpatRast} created using the create_base_raster() on which the
DEM data will be extracted, currently accepts raster (.tif) or filepath.}

\item{res}{resolution in meters of final project}

\item{out_dir}{the root directory to hold the cded dem spatRast file. If not
specified uses the default from the \code{fid} folder structure.}

\item{write_output}{should the snapped aoi bounding box be written to disk?
If \code{TRUE} (default), will write to \code{out_dir}}

\item{overwrite}{logical. If \code{TRUE}, \code{filename} is overwritten}
}
\value{
a spatRast file of CDED data
}
\description{
This function is a wrapper for bcmaps::cded_terra() to extract data from the
Canadian Digital Elevation Data set to match the raster template built using
create_raster_template().
}
\examples{
\dontrun{
get_cded_dem(aoi_bb = file.path(fid$shape_dir_1010[2],"aoi_snapped.gpkg"),
    res = 5, out_dir = fid$cov_dir_1020[2]))
}
}
