% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupfolders.R
\name{setupfolders}
\alias{setupfolders}
\title{Setup Folder Structure}
\usage{
setupfolders(aoi)
}
\arguments{
\item{aoi}{A character string of the folders}
}
\value{
a named list of the folders created
}
\description{
Sets up the default folder structure
}
\details{
Folder structure contains all the raw data, covariates generated, and
subsequent models, and predicted map rasters.
}
\examples{
## add more
setupfolder("AppleValley")
}
